Enable options that allow the higlass viewer to go full screen. (#43)
Specifically, a signal can be sent to turn on and off the various tools in the toolbar. If all tools are off, the entire left panel is not rendered.

Alternately, the left panel can now be collapsed and hidden from view.

Work incident to collapsing the side panel resulted in allowing us to remember certain critical data over refreshes.